Update assembler.js
Fixed monitoring enable bug.

The monitoring callback used to look at the first monitoring button in the page. This meant if you clicked on the third monitoring button in the page, the callback would then look at the state of the first monitoring button in the page. And it would thus look at the wrong button's state.